Feature: 랭킹 상세 조회 기능 추가 (#317)
* Feat: ErrorCode RANKING_NOT_FOUND 추가

* Feat: 상세 랭킹 조회 API 추가, 랭킹 조회 시 Redis 사용

* Feat: ë¡레벨업, 인증글 작성 시 퀘스트 진행도 업데이트 추가

* Feat: QuestConditionType 수정

* Fix: UserQuest 생성 시 progressCount가 증가된 상태로 생성되는 문제 해결

* Feat: 랭킹 검색 LOWER 삭제

* Docs: RankingControllerDocs 추가

* Refactor: RequestDto 어노테이션 위치 수정

+++ src/main/java/com/explorer/gabom/domain/level/service/LevelService.java
@@ -15,7 +15,7 @@ public class LevelService {
 
 	private final LevelRepository levelRepository;
 
-	public int calculateLevel(int currentExp) {
+	public int calculateLevel(Long currentExp) {
 		List<Level> levels = levelRepository.findAllByOrderByRequiredExpAsc();
 		int userLevel = 1;
