+++ src/main/java/com/explorer/gabom/domain/quest/dto/request/QuestUpdateRequest.java
@@ -29,11 +29,11 @@ public class QuestUpdateRequest {
 
 	@Min(value = 1, message = "포인트는 1 이상이어야 합니다.")
 	@Schema(description = "퀘스트 완료 시 지급할 포인트", example = "500", minimum = "1")
-	private Integer rewardPoint;
+	private Long rewardPoint;
 
 	@Min(value = 1, message = "경험치는 1 이상이어야 합니다.")
 	@Schema(description = "퀘스트 완료 시 지급할 경험치", example = "100", minimum = "1")
-	private Integer rewardExp;
+	private Long rewardExp;
 
 	@Schema(description = "퀘스트 완료 시 지급할 칭호 ID", example = "1")
 	private Long rewardTitleId;
